refactor(doctor): extract reusable report engine
